follow-up: move Idefics2 to a shared location to eliminate unexpected dependency. (#6603)
